pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "decoct"
version = "0.1.0"
description = "Infrastructure context compression for LLMs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "decoct contributors" },
]
keywords = ["llm", "yaml", "infrastructure", "context", "compression"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "ruamel.yaml>=0.18",
    "tiktoken>=0.7",
    "click>=8.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "defusedxml>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff>=0.8",
    "mypy>=1.0",
]
llm = [
    "anthropic>=0.40",
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.scripts]
decoct = "decoct.cli:cli"

[project.urls]
Homepage = "https://decoct.io"
Repository = "https://github.com/decoct-io/decoct"
Documentation = "https://decoct.io/docs"
Issues = "https://github.com/decoct-io/decoct/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/decoct"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
# Spec-mandated UPPER_CASE function names (§2.4, §7.3)
"src/decoct/core/canonical.py" = ["N802"]
"src/decoct/core/hashing.py" = ["N802"]
# Spec-mandated H_norm parameter name (§3.4)
"src/decoct/analysis/tier_classifier.py" = ["N803"]
# Scripts use sys.path.insert before imports
"scripts/*.py" = ["E402", "I001"]
# Archetypal fixtures use conftest-injected helpers module
"tests/fixtures/archetypal/*.py" = ["I001"]
"tests/fixtures/archetypal/integration/*.py" = ["I001"]
# Generator scripts provided as-is
"tests/fixtures/archetypal/integration/generate_integration.py" = ["E501", "F401", "F841"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
